Indexer Example 1
Indexers allow instances of a class or struct to be indexed just like arrays. The indexed value can be set or retrieved without explicitly specifying a type or instance member. Indexers resemble properties except that their accessors take parameters.

To declare an indexer on a class or struct, use the 'this' keyword.